+++ Lavender/Lavender.cs
@@ -1,11 +1,13 @@
-﻿using Lavender.FurnitureLib;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
 using UnityEngine;
 using HarmonyLib;
+using Lavender.FurnitureLib;
+using Lavender.ItemLib;
+using FullSerializer;
 
 namespace Lavender
 {
@@ -35,6 +37,7 @@ namespace Lavender
             harmony = new Harmony(LCMPluginInfo.PLUGIN_GUID);
 
             harmony.PatchAll(typeof(FurniturePatches));
+            harmony.PatchAll(typeof(ItemPatches));
         }
 
         #region FurnitureLib
@@ -134,6 +137,78 @@ namespace Lavender
             }
 
             return true;
+        }
+
+        #endregion
+
+        #region ItemLib
+
+        private static readonly fsSerializer JSON_serializer = new fsSerializer();
+
+        public static List<Item> customItemDatabase;
+
+        public static void AddCustomItem(Item item, string mod_name)
+        {
+            item.Categories.AddToArray("Lavender ModItem");
+            item.Categories.AddToArray(mod_name);
+
+            customItemDatabase.Add(item);
+        }
+
+        public static void AddCustomItemsFromJson(string jsonPath, string mod_name)
+        {
+            if (!File.Exists(jsonPath)) { LavenderLog.Error($"AddCustomItemsFromJson(): File at path '{jsonPath}' doesn't exists!");  return; }
+
+            try
+            {
+                fsData data = fsJsonParser.Parse(File.ReadAllText(jsonPath));
+                object result = null;
+                ItemDatabase.JSON_serializer.TryDeserialize(data, typeof(List<Item>), ref result).AssertSuccessWithoutWarnings();
+
+                List<Item> Items = result as List<Item>;
+
+                foreach (Item i in Items)
+                {
+                    string path = jsonPath.Substring(0, jsonPath.Length - Path.GetFileName(jsonPath).Length);
+
+                    // Sprite Path
+                    if(i.Appearance.SpritePath.EndsWith(".png") || i.Appearance.SpritePath.EndsWith(".jpg"))
+                    {
+                        i.Appearance.SpritePath = "Lavender_SRC#" + path + i.Appearance.SpritePath;
+                    }
+                    else if(i.Appearance.SpritePath.Contains("#AB"))
+                    {
+                        // path to assetbundle + #AB<Sprite_Name>
+                        i.Appearance.SpritePath = "Lavender_AB#" + path + i.Appearance.SpritePath;
+                    }
+
+                    // Prefab Path
+                    if(i.Appearance.PrefabPath.EndsWith(".obj"))
+                    {
+                        i.Appearance.PrefabPath = "Lavender_SRC#" + path + i.Appearance.PrefabPath;
+                    }
+                    else if (i.Appearance.PrefabPath.Contains("#AB")) 
+                    {
+                        i.Appearance.PrefabPath = "Lavender_AB#" + path + i.Appearance.PrefabPath;
+                    }
+
+                    // Prefab Path Many
+                    if (i.Appearance.PrefabPathMany.EndsWith(".obj"))
+                    {
+                        i.Appearance.PrefabPathMany = "Lavender_SRC#" + path + i.Appearance.PrefabPathMany;
+                    }
+                    else if (i.Appearance.PrefabPathMany.Contains("#AB"))
+                    {
+                        i.Appearance.PrefabPathMany = "Lavender_AB#" + path + i.Appearance.PrefabPathMany;
+                    }
+
+                    AddCustomItem(i, mod_name);
+                }
+            }
+            catch(Exception e)
+            {
+                LavenderLog.Error($"Error while loading '{mod_name}'s Item Database!\nException: {e}");
+            }
         }
 
         #endregion